add cli args, group noops, show unknown opcodes
The bunk of the changes here is that the disassembler now takes a command line input for which file to disassemble.

I've also made it so that instead of crashing, the disassembler keeps track of how many times each invalid instruction shows up. NOP instructions were grouped together as well.